fixed casting

+++ ex01/Span.cpp
@@ -14,8 +14,10 @@ Span::Span(Span const &other)
       _numbers(other._numbers),
       _sortedNumbers(other._sortedNumbers) {}
 
-Span &Span::operator=(Span const &other) {
-    if (this != &other) {
+Span &Span::operator=(Span const &other)
+{
+    if (this != &other)
+    {
         _maxSize = other._maxSize;
         _numbers = other._numbers;
         _sortedNumbers = other._sortedNumbers;
@@ -24,20 +26,20 @@ Span &Span::operator=(Span const &other) {
 }
 
 // public methods
-void Span::addNumber(int number) {
-    if (_numbers.size() >= _maxSize) {
+void Span::addNumber(int number)
+{
+    if (_numbers.size() >= _maxSize)
+    {
         throw std::out_of_range("Span is full");
     }
     _numbers.push_back(number);
     _sortedNumbers.insert(number);
 }
 
-// static unsigned int min(unsigned int a, unsigned int b) {
-//     return (a < b) ? a : b;
-// }
-
-unsigned int Span::shortestSpan() const {
-    if (_numbers.size() < 2) {
+unsigned int Span::shortestSpan() const
+{
+    if (_numbers.size() < 2)
+    {
         throw std::logic_error("Not enough numbers to find a span");
     }
 
@@ -46,12 +48,15 @@ unsigned int Span::shortestSpan() const {
     ++next;
 
     unsigned int minSpan = UINT_MAX;
-    while (next != _sortedNumbers.end()) {
-        unsigned int span = static_cast<unsigned int>(*next - *current);
-        if (span < minSpan) {
+    while (next != _sortedNumbers.end())
+    {
+        unsigned int span = static_cast<unsigned int>(static_cast<long>(*next) - static_cast<long>(*current));
+        if (span < minSpan)
+        {
             minSpan = span;
         }
-        if (minSpan == 0) {
+        if (minSpan == 0)
+        {
             return 0;
         }
         ++current;
@@ -60,13 +65,15 @@ unsigned int Span::shortestSpan() const {
     return minSpan;
 }
 
-unsigned int Span::longestSpan() const {
-    if (_numbers.size() < 2) {
+unsigned int Span::longestSpan() const
+{
+    if (_numbers.size() < 2)
+    {
         throw std::logic_error("Not enough numbers to find a span");
     }
     int minNum = *std::min_element(_numbers.begin(), _numbers.end());
     int maxNum = *std::max_element(_numbers.begin(), _numbers.end());
-    return maxNum - minNum;
+    return static_cast<unsigned int>(static_cast<long>(maxNum) - static_cast<long>(minNum));
 }
 
 size_t Span::getSize() const { return _numbers.size(); }
